completed the card and dice in class assignment and added it to the portfolio
found another bug with solitair being the win conidtion if cards were
removed from the winning pile and placed back it still incremented the
count and thus was greater than indtended by the end of the game. I
tried counting the number of cards but I don't know if it will work and
I don't have time to test.